+++ KTW_DX/GameEngineContents/ContentsSoundPlayer.cpp
@@ -19,8 +19,8 @@ void ContentsSoundPlayer::RandomSoundPlay()
 	int SoundListSize = static_cast<int>(SoundList.size());
 
 	GameEngineRandom Random;
-	unsigned int Time = static_cast<unsigned int>(time(NULL));
-	Random.SetSeed(static_cast<long long>(Time));
+	long long Time = static_cast<long long>(time(NULL));
+	Random.SetSeed(Time);
 
 	int SoundNum = Random.RandomInt(0, SoundListSize - 1);
