+++ KTW_DX/GameEngineContents/DogFish_State.cpp
@@ -92,8 +92,8 @@ void DogFish::MoveUpdate(float _Delta)
 void DogFish::DeathSoundPlay()
 {
 	GameEngineRandom Random;
-	unsigned int Time = static_cast<unsigned int>(time(NULL));
-	Random.SetSeed(static_cast<long long>(Time));
+	long long Time = static_cast<long long>(time(NULL));
+	Random.SetSeed(Time);
 
 	int SoundNum = Random.RandomInt(0, 1);
