+++ KTW_DX/GameEngineContents/FirstBossPhase3_Sound.cpp
@@ -4,8 +4,8 @@
 void FirstBossPhase3::IdleSoundPlay()
 {
 	GameEngineRandom Random;
-	unsigned int Time = static_cast<unsigned int>(time(NULL));
-	Random.SetSeed(static_cast<long long>(Time));
+	long long Time = static_cast<long long>(time(NULL));
+	Random.SetSeed(Time);
 
 	int SoundNum = Random.RandomInt(0, 3);
 
@@ -31,8 +31,8 @@ void FirstBossPhase3::MoveSoundPlay()
 	MoveSound.Stop();
 
 	GameEngineRandom Random;
-	unsigned int Time = static_cast<unsigned int>(time(NULL));
-	Random.SetSeed(static_cast<long long>(Time));
+	long long Time = static_cast<long long>(time(NULL));
+	Random.SetSeed(Time);
 
 	int SoundNum = Random.RandomInt(0, 4);
 
@@ -59,8 +59,8 @@ void FirstBossPhase3::MoveSoundPlay()
 void FirstBossPhase3::MoveVoiceSoundPlay()
 {
 	GameEngineRandom Random;
-	unsigned int Time = static_cast<unsigned int>(time(NULL));
-	Random.SetSeed(static_cast<long long>(Time));
+	long long Time = static_cast<long long>(time(NULL));
+	Random.SetSeed(Time);
 
 	int SoundNum = Random.RandomInt(0, 3);
 
@@ -84,8 +84,8 @@ void FirstBossPhase3::MoveVoiceSoundPlay()
 void FirstBossPhase3::AttackSoundPlay()
 {
 	GameEngineRandom Random;
-	unsigned int Time = static_cast<unsigned int>(time(NULL));
-	Random.SetSeed(static_cast<long long>(Time));
+	long long Time = static_cast<long long>(time(NULL));
+	Random.SetSeed(Time);
 
 	int SoundNum = Random.RandomInt(0, 4);
