+++ KTW_DX/GameEngineContents/FirstBoss_Phase1_Sound.cpp
@@ -4,8 +4,8 @@
 void FirstBoss::Phase1MoveSoundPlay()
 {
 	GameEngineRandom Random;
-	unsigned int Time = static_cast<unsigned int>(time(NULL));
-	Random.SetSeed(static_cast<long long>(Time));
+	long long Time = static_cast<long long>(time(NULL));
+	Random.SetSeed(Time);
 
 	int SoundNum = Random.RandomInt(0, 3);
 
@@ -29,8 +29,8 @@ void FirstBoss::Phase1MoveSoundPlay()
 void FirstBoss::Phase1LandSoundPlay()
 {
 	GameEngineRandom Random;
-	unsigned int Time = static_cast<unsigned int>(time(NULL));
-	Random.SetSeed(static_cast<long long>(Time));
+	long long Time = static_cast<long long>(time(NULL));
+	Random.SetSeed(Time);
 
 	int SoundNum = Random.RandomInt(0, 4);
 
@@ -57,8 +57,8 @@ void FirstBoss::Phase1LandSoundPlay()
 void FirstBoss::Phase1AttackSoundPlay()
 {
 	GameEngineRandom Random;
-	unsigned int Time = static_cast<unsigned int>(time(NULL));
-	Random.SetSeed(static_cast<long long>(Time));
+	long long Time = static_cast<long long>(time(NULL));
+	Random.SetSeed(Time);
 
 	int SoundNum = Random.RandomInt(0, 1);
