+++ KTW_DX/GameEngineContents/Ram_Arm.cpp
@@ -152,8 +152,8 @@ bool Ram_Arm::DeathPosCheck()
 void Ram_Arm::ClapSoundPlay()
 {
 	GameEngineRandom Random;
-	unsigned int Time = static_cast<unsigned int>(time(NULL));
-	Random.SetSeed(static_cast<long long>(Time));
+	long long Time = static_cast<long long>(time(NULL));
+	Random.SetSeed(Time);
 
 	int SoundNum = Random.RandomInt(0, 3);
