+++ KTW_DX/GameEngineContents/ShipBoss.cpp
@@ -300,8 +300,8 @@ void ShipBoss::CreateBeam()
 void ShipBoss::BubbleShootSoundPlay()
 {
 	GameEngineRandom Random;
-	unsigned int Time = static_cast<unsigned int>(time(NULL));
-	Random.SetSeed(static_cast<long long>(Time));
+	long long Time = static_cast<long long>(time(NULL));
+	Random.SetSeed(Time);
 
 	int SoundNum = Random.RandomInt(0, 3);
