+++ KTW_DX/GameEngineContents/Spider_Head.cpp
@@ -163,8 +163,8 @@ void Spider_Head::LandCountCheck()
 void Spider_Head::FallSoundPlay()
 {
 	GameEngineRandom Random;
-	unsigned int Time = static_cast<unsigned int>(time(NULL));
-	Random.SetSeed(static_cast<long long>(Time));
+	long long Time = static_cast<long long>(time(NULL));
+	Random.SetSeed(Time);
 
 	int SoundNum = Random.RandomInt(0, 5);
 
@@ -194,8 +194,8 @@ void Spider_Head::FallSoundPlay()
 void Spider_Head::LandSoundPlay()
 {
 	GameEngineRandom Random;
-	unsigned int Time = static_cast<unsigned int>(time(NULL));
-	Random.SetSeed(static_cast<long long>(Time));
+	long long Time = static_cast<long long>(time(NULL));
+	Random.SetSeed(Time);
 
 	int SoundNum = Random.RandomInt(0, 5);
