+++ KTW_DX/GameEngineContents/SummonAttackObject.cpp
@@ -49,8 +49,8 @@ bool SummonAttackObject::CameraOutCheck()
 void SummonAttackObject::AttackSoundPlay()
 {
 	GameEngineRandom Random;
-	unsigned int Time = static_cast<unsigned int>(time(NULL));
-	Random.SetSeed(static_cast<long long>(Time));
+	long long Time = static_cast<long long>(time(NULL));
+	Random.SetSeed(Time);
 
 	int SoundNum = Random.RandomInt(0, 3);
