+++ KTW_DX/GameEngineContents/SummonDevil.cpp
@@ -230,8 +230,8 @@ void SummonDevil::HitCheck()
 void SummonDevil::SpawnSoundPlay()
 {
 	GameEngineRandom Random;
-	unsigned int Time = static_cast<unsigned int>(time(NULL));
-	Random.SetSeed(static_cast<long long>(Time));
+	long long Time = static_cast<long long>(time(NULL));
+	Random.SetSeed(Time);
 
 	int SoundNum = Random.RandomInt(0, 2);
 
@@ -252,8 +252,8 @@ void SummonDevil::SpawnSoundPlay()
 void SummonDevil::DeathSoundPlay()
 {
 	GameEngineRandom Random;
-	unsigned int Time = static_cast<unsigned int>(time(NULL));
-	Random.SetSeed(static_cast<long long>(Time));
+	long long Time = static_cast<long long>(time(NULL));
+	Random.SetSeed(Time);
 
 	int SoundNum = Random.RandomInt(0, 3);
